refactor(view): create racing.view pkg
- RacingGameStarter: View 라는 이름이 안 어울려 Starter 로 이름 변경
