+++ src/main/java/racing/RacingGameStarter.java
@@ -1,8 +1,11 @@
 package racing;
 
+import racing.view.InputView;
+import racing.view.RacingNotifyView;
+
 import java.util.List;
 
-public class RacingGameView {
+public class RacingGameStarter {
 
 	public static void main(String[] args) {
 		InputView inputView = new InputView();
