+++ src/main/java/racing/view/InputView.java
@@ -1,14 +1,14 @@
-package racing;
+package racing.view;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 
-class InputView {
+public class InputView {
 
 	private final Scanner scanner = new Scanner(System.in);
 
-	List<String> inputCarNameList() {
+	public List<String> inputCarNameList() {
 		String carNameInput;
 		do {
 			System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분");
@@ -42,7 +42,7 @@ class InputView {
 		}
 	}
 
-	int inputRaceTurn() {
+	public int inputRaceTurn() {
 		System.out.println("시도할 회 수는 몇 회 인가요?");
 		return this.scanner.nextInt();
 	}
